move test_stubs and test_samples to pytest
